Implemented live edit of script sources.
Introduced `callbackForNextBreak` that allows us to handle break after
'step in' command within back-end.

Renamed changeConsolePanel to changeConsole and changed it to show
console as popup instead of switching to Console panel. The reason
is that when there is an error while saving live-edit changes, front-end
switches immediately back to Sources panel.