Changed fetchCallFrames to not fetch scope details
Modified CallFramesProvider so that it does not fetch scope details
together with call frames. It creates a special objectId instead.

RuntimeAgent.getProperties checks for this special objectId format
and delegates getProperties to CallFramesProvider, which sends
'scope' debugger request to get scope details.

This change should reduce the time between debugged process hitting
a breakpoint and node-inspector pausing in the browser.