ScriptManager: fixed crash (mocha --debug-brk)
Modified ScriptManager.addScript to immediately add the script in the internal
registry and load sourcemap later.

This change fixed a crash when debugging mocha unit-tests, where 'break' event
arrived immediately after 'afterCompile' and node-inspector could not find
the script where the program stopped, because registration was wating for
for 'scripts' response from debugger.