Added `stackTraceLimit` configuration, fixes #96.
The new option `stackTraceLimit`, like the corresponding option in
Chrome, allows us to specify how many stack frames we want when
stopped at a breakpoint. It can be set as high as you like.
Negative numbers will cause an error.

The default number of frames is set to 50 to prevent possible
performance issues; if you wish to capture all frames, no matter the
consequences, set this option to a high number, like `99999`.