Make sure to open URL in a webkit browser
`opener` always opens the URL in the default browser of the user.
It might be annoying for a user whose default browser is Firefox
to have Node Inspector start in Firefox.

This PR makes sure that the Node Inspector URL is opened
either in Chrome, Chromium or Opera (if one of them is the default
browsers, it gets priority). If we can't do that for some reason
(e.g. user has none of them installed), we just display an error
message in the console but do not open the URL in any browser.

Close #346
Close #528