Update README.md
Add exclusive include to make sure people know how they can include only certain files. This is extremely useful if you're trying to debug a certain file.

I also changed one of the examples on line 340, because the options have to be specific before the main script, or they will not be included.