Change node-inspector to support the remote debugging.
:Detailed Notes:
node-inspector had to run on the same machine
with the debuggable nodejs process.
This patch supports a remote debugging for the debuggable nodejs
running on the difference machine via ip address.

:Example:
./bin/inspector.js --debug-host='192.168.0.12' --debug-port=5858
or put the host address and port in the config.json and run inspector.js
or can change the value of the host argument in the url.
  ex) http://127.0.0.1:8080/debug?host=192.168.0.12&port=5858

:Etc:
- This patch aims to merge into node-inspector v0.7.x branch.
- As KeyboardEvent.keyIdentifier has been removed
  from thre recent browsers including chrome 54 rev,
  front-end should use KeyboardEvent.key instead.